Put `renderless` code in `render` module
