+++ src/lib.rs
@@ -42,5 +42,4 @@ pub mod nodes;
 pub mod params;
 pub mod puppet;
 pub mod render;
-pub mod renderless;
 pub mod texture;
