+++ src/params.rs
@@ -6,7 +6,7 @@ use crate::math::interp::{
 use crate::math::matrix::Matrix2d;
 use crate::nodes::node::InoxNodeUuid;
 use crate::puppet::Puppet;
-use crate::renderless::{RenderCtxKind, NodeRenderCtxs, PartRenderCtx};
+use crate::render::{RenderCtxKind, NodeRenderCtxs, PartRenderCtx};
 
 /// Parameter binding to a node. This allows to animate a node based on the value of the parameter that owns it.
 #[derive(Debug, Clone)]
